Unified get() methods in objects()
The scene don't need to known what the GameObjects it contain implement or inherit.
Creating a single objects() based on typecast allow to remove dependencies and the need to update the class if the GameObject class is updated.
Note that the type casting is done without checking as we don't known to what we want to cast.